% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_agreement.R
\name{fs_agreement}
\alias{fs_agreement}
\title{Calcula el acuerdo entre métodos de selección de características}
\usage{
fs_agreement(ml_fs_obj)
}
\arguments{
\item{ml_fs_obj}{Un objeto de clase \code{ml_fs} que contiene los resultados de la selección de características.}
}
\value{
Un data frame con:
\itemize{
\item Una fila por cada característica seleccionada por al menos un método.
\item Columnas lógicas indicando qué métodos seleccionaron la característica.
\item Una columna \code{agree} con el número total de métodos que seleccionaron esa característica.
}
}
\description{
Esta función evalúa cuántos métodos de selección han seleccionado cada característica.
Devuelve un data frame donde cada fila representa una característica, con columnas indicando
si fue seleccionada por cada método y cuántos métodos la seleccionaron en total.
}
