% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_cv.R
\name{get_best_cv}
\alias{get_best_cv}
\title{Obtener resumen de mejores métricas CV por combinación}
\usage{
get_best_cv(ml_cs_object)
}
\arguments{
\item{ml_cs_object}{Objeto de clase \code{ml_cs} generado tras evaluación.}
}
\value{
Tibble ordenado con combinaciones y su métrica CV.
}
\description{
Extrae y ordena el desempeño de validación cruzada según la métrica seleccionada.
}
