% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_test.R
\name{get_best_test}
\alias{get_best_test}
\title{Obtener mejores métricas de test por combinación}
\usage{
get_best_test(ml_cs_object)
}
\arguments{
\item{ml_cs_object}{Objeto de clase \code{ml_cs} que contiene resultados de evaluación.}
}
\value{
Tibble ordenado con combinaciones y su métrica en test.
}
\description{
Extrae y ordena el desempeño en datos de test según la métrica seleccionada.
}
