% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_metrics.R
\name{get_test_metrics}
\alias{get_test_metrics}
\title{Obtener métricas de test por combinación FS + modelo}
\usage{
get_test_metrics(object, metodoFS, modelo, ...)
}
\arguments{
\item{object}{Objeto de clase \code{ml_cs} que contiene los resultados del proceso de clasificación.}

\item{metodoFS}{Nombre del método de selección de características utilizado.}

\item{modelo}{Nombre del modelo evaluado.}

\item{...}{Argumentos adicionales (no utilizados actualmente).}
}
\value{
Un data frame con las métricas de evaluación en test para la combinación especificada.
}
\description{
Esta función extrae las métricas de evaluación en el conjunto de test
para una combinación específica de método de selección de características y modelo.
}
