% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_metrics.R
\name{get_training_metrics}
\alias{get_training_metrics}
\title{Obtener métricas de entrenamiento por combinación FS + modelo}
\usage{
get_training_metrics(object, metodoFS, modelo)
}
\arguments{
\item{object}{Objeto de clase \code{ml_cs} que contiene los resultados del proceso de clasificación.}

\item{metodoFS}{Nombre del método de selección de características utilizado.}

\item{modelo}{Nombre del modelo evaluado.}
}
\value{
Un data frame con las métricas de validación cruzada para la combinación especificada.
}
\description{
Esta función permite extraer las métricas de validación cruzada obtenidas
durante el entrenamiento de un modelo específico, asociado a un método de selección de características.
}
