% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_preprocessor.R
\name{ml_preprocessor}
\alias{ml_preprocessor}
\title{Crear objeto ml_preprocessor con paralelización}
\usage{
ml_preprocessor(data, target, seed = 99, workers = 2, verbose = TRUE)
}
\arguments{
\item{data}{Data frame con los datos originales.}

\item{target}{Nombre de la variable objetivo (como string).}

\item{seed}{Semilla para la aleatorización.}

\item{workers}{Número de núcleos o workers a usar para la paralelización.}

\item{verbose}{Mensajes de progreso.}
}
\value{
Un objeto de clase \code{ml_preprocessor} con los datos procesados y divididos.
}
\description{
Esta función preprocesa los datos, divide en entrenamiento/test y devuelve un objeto de clase \code{ml_preprocessor}.
Utiliza paralelización automática con \code{future} para acelerar el preprocesamiento.
}
