% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_cv.R
\name{plot_metric_cv}
\alias{plot_metric_cv}
\title{Graficar una métrica de validación cruzada por modelo y método FS}
\usage{
plot_metric_cv(object, metrica)
}
\arguments{
\item{object}{Objeto de clase \code{ml_cs} que contiene los resultados de validación cruzada.}

\item{metrica}{Nombre de la métrica a graficar (por ejemplo, \code{"accuracy"}, \code{"roc_auc"}).}
}
\value{
Un objeto \code{ggplot} con la visualización de la métrica seleccionada.
}
\description{
Esta función genera una gráfica de barras con errores estándar para una métrica de validación cruzada
(por ejemplo, accuracy, ROC AUC, etc.) según el modelo y el método de selección de características.
}
