% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_test.R
\name{plot_metric_test}
\alias{plot_metric_test}
\title{Plot de métricas de test para modelos de clasificación}
\usage{
plot_metric_test(object, metrica)
}
\arguments{
\item{object}{Un objeto de clase \code{ml_cs}, generado por la función \code{ml_cs()}.}

\item{metrica}{Cadena de texto que indica la métrica de evaluación a graficar.}
}
\value{
Un gráfico de ggplot2 con la métrica de test para cada modelo y método FS.
}
\description{
Genera un gráfico de barras con los valores de una métrica de test para
cada combinación de método de selección de características y modelo.
}
