% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Graficar la curva ROC para una combinación de modelo y método de selección}
\usage{
plot_roc(ml_cs_object, metodo_fs, modelo)
}
\arguments{
\item{ml_cs_object}{Objeto de clase \code{ml_cs} generado por la función de evaluación.}

\item{metodo_fs}{Nombre del método de selección de características utilizado.}

\item{modelo}{Nombre del modelo de clasificación evaluado.}
}
\value{
Un objeto \code{ggplot} con la curva ROC.
}
\description{
Esta función genera la curva ROC para una combinación específica de método de selección de características y modelo,
utilizando los datos almacenados en un objeto de clase \code{ml_cs}.
}
