% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_times.R
\name{plot_times}
\alias{plot_times}
\title{Gráfico del tiempo total por combinación de FS + modelo}
\usage{
plot_times(ml_fs, ml_cs)
}
\arguments{
\item{ml_fs}{Objeto de clase \code{ml_fs} con los tiempos de selección de características.}

\item{ml_cs}{Objeto de clase \code{ml_cs} con los tiempos de entrenamiento de modelos.}
}
\value{
Un gráfico de barras (\code{ggplot}) con los tiempos totales por combinación.
}
\description{
Esta función visualiza el tiempo total de computación por combinación de método
de selección de características (FS) y modelo de clasificación, sumando los tiempos
de la fase de selección y del entrenamiento del modelo.
}
