% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ml_cs.R
\name{summary.ml_cs}
\alias{summary.ml_cs}
\title{Resumen de un objeto de clase \code{ml_cs}}
\usage{
\method{summary}{ml_cs}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{ml_cs} devuelto por la función \code{ml_cs2()}.}

\item{...}{Argumentos adicionales (no utilizados actualmente).}
}
\value{
Imprime un resumen en consola.
}
\description{
Muestra un resumen del proceso de evaluación de modelos, incluyendo
los métodos de selección de características utilizados, los modelos evaluados,
el número total de combinaciones evaluadas y los tiempos de entrenamiento.
}
