% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ml_fs.R
\name{summary.ml_fs}
\alias{summary.ml_fs}
\title{Resumen del objeto \code{ml_fs}}
\usage{
\method{summary}{ml_fs}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{ml_fs}.}
}
\value{
Un data frame con el resumen .
}
\description{
Devuelve un resumen con el número de variables seleccionadas y el tiempo de ejecución para cada método de selección de características.
}
