% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ml.R
\name{summary.ml_preprocessor}
\alias{summary.ml_preprocessor}
\title{Resumen de objeto ml_preprocessor}
\usage{
\method{summary}{ml_preprocessor}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{ml_preprocessor}.}
}
\description{
Muestra un resumen básico del objeto \code{ml_preprocessor}, incluyendo
la variable objetivo, tamaño y distribución de clases en los conjuntos
de entrenamiento y test.
}
