% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fs.R
\name{time_fs}
\alias{time_fs}
\title{Gráfico de tiempos de ejecución de métodos de FS}
\usage{
time_fs(ml_fs)
}
\arguments{
\item{ml_fs}{Objeto de clase \code{ml_fs}.}
}
\value{
Un objeto \code{ggplot}.
}
\description{
Genera un gráfico de barras con los tiempos (en segundos) que tardó cada método de selección.
}
