% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_features.R
\name{venn_features}
\alias{venn_features}
\title{Diagrama de Venn para visualizar FS}
\usage{
venn_features(ml_fs, methods)
}
\arguments{
\item{ml_fs}{Objeto de clase \code{ml_fs}.}

\item{methods}{Vector de nombres de métodos a incluir en el diagrama (mínimo 2, máximo 5).}
}
\value{
Un objeto \code{ggplot} con el diagrama de Venn.
}
\description{
Genera un diagrama de Venn usando el paquete \code{ggvenn} para visualizar la intersección
de variables seleccionadas por distintos métodos de selección de características.
}
